pyproject.toml
[project]
name = "deepagent-document-generation"
version = "0.1.0"
description = "Multi-agent async pipeline for BRD generation using deepagents library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.1.5",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.0.0",
    "polars>=0.19.0",
    "openpyxl>=3.1.0",
    "tiktoken>=0.5.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]